# config/swagger.json
{
    "swagger": "2.0",
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "info": {
        "title": "FLOW CARDIT API",
        "version": "1.0"
    },
    "paths": {
        "/users": {
            "get": {
                "description": "Index of users",
                "responses": {
                    "200": {
                        "description": "Returns a list of all users",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    },
                    "400": {
                        "description": "Failed to show users"
                    }
                }
            },
            "post": {
                "description": "Creates new user",
                "parameters": [
                    {
                        "$ref": "#/parameters/userData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New user created",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Failed to create user"
                    },
                    "500": {
                        "description": "Failed to hash password"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Logs in user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": "true",
                        "type": "object",
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs in user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Auth failed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/login/password": {
            "put": {
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "required": "true"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Show user by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user by id",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "User with given id does not exist"
                    }
                }
            },
            "put": {
                "description": "Update user by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    },
                    {
                        "$ref": "#/parameters/userData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Failed to update user"
                    },
                    "500": {
                        "description": "Failed to hash password"
                    }
                }
            },
            "delete": {
                "description": "Delete user by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "User with given id does not exist."
                    }
                }
            }
        },
        "/users/{id}/me": {
            "get": {
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ]
            },
            "put": {
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "properties": {
                                "password": {
                                    "type": "string",
                                    "required": "true"
                                }
                            }
                        }
                    }
                ]
            }
        },
...
        "/groups/{groupId}/users": {
            "get": {
                "description": "Show users by groupId",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show users by groupId"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "LoginRequest": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "User": {
            "properties": {
                "firstName": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                },
                "lastName": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                },
                "email": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "defaultValue": null,
                    "unique": true,
                    "validate": {
                        "len": [
                            3,
                            500
                        ],
                        "isEmail": true
                    }
                },
                "passwordHash": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "defaultValue": null,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "contributor",
                        "student"
                    ],
                    "allowNull": false,
                    "notEmpty": true,
                    "defaultValue": null,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                },
                "lastLogin": {
                    "type": "string",
                    "isDate": true,
                    "defaultValue": "new Date()"
                },
                "GroupId": {
                    "type": "integer"
                }
            }
        },
        "Users": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/User"
            }
        },
...
    },
    "parameters": {
        "id": {
            "name": "id",
            "in": "path",
            "required": "true",
            "type": "integer"
        },
...
        "userData": {
            "name": "user",
            "in": "body",
            "schema": {
                "$ref": "#/definitions/User"
            }
        },
...
    }
}